refactor(packager): move packager compile logic to a electron-packager afterCopy hook
ISSUES CLOSED: Fixes #7
