+++ package.json
@@ -52,7 +52,6 @@
     ]
   },
   "dependencies": {
-    "asar": "^0.12.3",
     "babel-register": "^6.16.3",
     "bluebird": "^3.4.6",
     "colors": "^1.1.2",
